add CategorySelector refactor at products modals.
