+++ client/src/components/pages/Store/locales/en.js
@@ -434,6 +434,7 @@ const storeEn = {
     title: "Categories",
     subtitle: "Manage your product categories",
     addCategory: "Add Category",
+    tableAriaLabel: "Categories table",
     name: "Name",
     type: "Type",
     actions: "Actions",
@@ -462,6 +463,7 @@ const storeEn = {
     title: "Products",
     subtitle: "Manage your store products",
     addProduct: "Add Product",
+    tableAriaLabel: "Products table",
     image: "Image",
     name: "Name",
     description: "Description",
