+++ client/src/components/pages/Store/locales/es.js
@@ -434,6 +434,7 @@ const storeEs = {
     title: "Categorías",
     subtitle: "Gestiona las categorías de tus productos",
     addCategory: "Agregar Categoría",
+    tableAriaLabel: "Tabla de categorías",
     name: "Nombre",
     type: "Tipo",
     actions: "Acciones",
@@ -462,6 +463,7 @@ const storeEs = {
     title: "Productos",
     subtitle: "Gestiona tu catálogo de productos e inventario",
     addProduct: "Agregar Producto",
+    tableAriaLabel: "Tabla de productos",
     image: "Imagen",
     name: "Nombre",
     description: "Descripción",
